Добавила абстрактный класс для компонентов. Разбила модуль util.js на несколько: utils/data, utils/common, utils/render
