+++ src/components/events-empty.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util';
+import AbstractComponent from './abstract-component';
 
 const createNoEventsTemplate = () => {
   return (
@@ -6,24 +6,8 @@ const createNoEventsTemplate = () => {
   );
 };
 
-export default class EventsEmptyComponent {
-  constructor() {
-    this._element = null;
-  }
-
+export default class EventsEmpty extends AbstractComponent {
   getTemplate() {
     return createNoEventsTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
