+++ src/components/info.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util';
+import AbstractComponent from './abstract-component';
 
 const createInfoTemplate = () => {
   return (
@@ -6,24 +6,8 @@ const createInfoTemplate = () => {
   );
 };
 
-export default class InfoComponent {
-  constructor() {
-    this._element = null;
-  }
-
+export default class Info extends AbstractComponent {
   getTemplate() {
     return createInfoTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
